refactor(receipts): extract core receipt schema microcrate
