+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
   "crates/bitnet-prompt-templates",
   "crates/bitnet-honest-compute",
   "crates/bitnet-receipts",
+  "crates/bitnet-receipts-core",
   "crates/bitnet-sampling",
       "crates/bitnet-transformer",  # Extracted from bitnet-models; depends on bitnet-quantization
   "crates/bitnet-device-probe",  # SRP: device detection / capability probing
@@ -95,6 +96,7 @@ default-members = [
   "crates/bitnet-inference",
   "crates/bitnet-prompt-templates",
   "crates/bitnet-receipts",
+  "crates/bitnet-receipts-core",
   "crates/bitnet-sampling",
   "crates/bitnet-cli",
   "crates/bitnet-server",
