+++ crates/bitnet-receipts-core/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "bitnet-receipts-core"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Core inference receipt schema, generation, and validation for BitNet"
+categories.workspace = true
+keywords.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+chrono = { workspace = true }
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-honest-compute = { path = "../bitnet-honest-compute", version = "0.2.1-dev" }
+bitnet-kernels = { path = "../bitnet-kernels", version = "0.2.1-dev", optional = true }
+rustc_version_runtime = "0.3.0"
+
+[features]
+gpu = ["dep:bitnet-kernels", "bitnet-kernels/gpu"]
+cuda = ["dep:bitnet-kernels", "bitnet-kernels/cuda"]
+
+[dev-dependencies]
+proptest = { workspace = true }
+insta = { workspace = true }
+tempfile = { workspace = true }
+serde_json = { workspace = true }
